Restructured all code and create new components for container of todos and nav bar
