+++ todo/src/components/NavBar.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StylesNavBar = styled.header`
+    width: 20%;
+    height: 100vh;
+    border: 3px solid red;
+
+`;
+
+export default function NavBar() {
+    return (
+        <StylesNavBar>
+            <nav>
+
+            </nav>
+        </StylesNavBar>
+    )
+}
